+++ graphql-maven-plugin-logic/src/main/java/com/graphql_java_generator/plugin/DocumentParser.java
@@ -721,7 +721,8 @@ public class DocumentParser {
 	void addTypeAnnotationForServerMode(Type o) {
 
 		if (!o.isInputType()) {
-			if (o instanceof ObjectType && !(o instanceof InterfaceType)) {
+			if (pluginConfiguration.getGenerateJPAAnnotation() && o instanceof ObjectType
+					&& !(o instanceof InterfaceType)) {
 				((AbstractType) o).addAnnotation("@Entity");
 			}
 		}
@@ -765,7 +766,7 @@ public class DocumentParser {
 	 * @param field
 	 */
 	void addFieldAnnotationForServerMode(Field field) {
-		if (!field.getOwningType().isInputType()) {
+		if (pluginConfiguration.getGenerateJPAAnnotation() && !field.getOwningType().isInputType()) {
 			if (field.isId()) {
 				// We have found the identifier
 				((FieldImpl) field).addAnnotation("@Id");