+++ graphql-maven-plugin-logic/src/main/resources/templates/client_query_target_type.vm.java
@@ -1,4 +1,4 @@
-package ${package};
+package ${pluginConfiguration.packageName};
 
 import java.util.List;