+++ graphql-maven-plugin-logic/src/main/resources/templates/server_GraphQLServerMain.vm.java
@@ -1,4 +1,4 @@
-package ${package};
+package ${pluginConfiguration.packageName};
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;