+++ graphql-maven-plugin-logic/src/main/resources/templates/server_GraphQLUtil.vm.java
@@ -1,4 +1,4 @@
-package ${package};
+package ${pluginConfiguration.packageName};
 
 import java.util.ArrayList;
 import java.util.List;