+++ graphql-maven-plugin-logic/src/test/java/com/graphql_java_generator/plugin/test/helper/PluginConfigurationTestHelper.java
@@ -35,4 +35,9 @@ public class PluginConfigurationTestHelper implements PluginConfiguration {
 	public PluginConfigurationTestHelper(Object caller) {
 		log = new Log4jLogger(caller);
 	}
+
+	@Override
+	public boolean getGenerateJPAAnnotation() {
+		return true;
+	}
 }