+++ graphql-maven-plugin/src/main/java/com/graphql_java_generator/mavenplugin/PluginConfigurationImpl.java
@@ -33,6 +33,11 @@ public class PluginConfigurationImpl implements PluginConfiguration {
 	}
 
 	@Override
+	public boolean getGenerateJPAAnnotation() {
+		return mojo.generateJPAAnnotation;
+	}
+
+	@Override
 	public Logger getLog() {
 		return log;
 	}